bothub/requirements.txt
flask
feedparser
requests
markdown
python-dotenv
python-telegram-bot
